rewrite decimals
