+++ Sourcecode/private/mx/core/Decimals.cpp
@@ -3,7 +3,6 @@
 // Distributed under the MIT License
 
 #include "mx/core/Decimals.h"
-#include "mx/core/PreciseDecimal.h"
 #include "mx/core/UnusedParameter.h"
 
 #include <sstream>
@@ -14,173 +13,287 @@ namespace mx
 {
     namespace core
     {
-        class Decimal::impl
+        PreciseDecimal::PreciseDecimal( short inMaxIntegerDigits, short inMaxDecimalDigits, DecimalType inValue )
+                : myIsNegative{ inValue < 0.0 }
+                , myInteger{ 0 }
+                , myDecimal{ 0 }
+                , myMaxIntegerDigits{ DEFAULT_MAX_INTEGER_DIGITS }
+                , myMaxDecimalDigits{ DEFAULT_MAX_DECIMAL_DIGITS }
         {
-        public:
-            impl( const DecimalType inValue )
-            : myValue( nullptr )
+            setMaxIntegerDigits( inMaxIntegerDigits );
+            setMaxDecimalDigits( inMaxDecimalDigits );
+            setMaxExpressibleNumber();
+            setValue( inValue );
+        }
+
+        PreciseDecimal::PreciseDecimal( short inMaxIntegerDigits, short inMaxDecimalDigits )
+        : PreciseDecimal{ inMaxIntegerDigits, inMaxDecimalDigits, 0.0 }
+        {
+
+        }
+
+        PreciseDecimal::PreciseDecimal( DecimalType inValue )
+        : PreciseDecimal{ DEFAULT_MAX_INTEGER_DIGITS, DEFAULT_MAX_DECIMAL_DIGITS, inValue }
+        {
+
+        }
+
+        PreciseDecimal::PreciseDecimal()
+        : PreciseDecimal{ DEFAULT_MAX_INTEGER_DIGITS, DEFAULT_MAX_DECIMAL_DIGITS, 0.0 }
+        {
+
+        }
+
+        DecimalType PreciseDecimal::getValue() const
+        {
+            return express( myInteger, myDecimal, myMaxDecimalDigits, myIsNegative );
+        }
+
+        static constexpr const long double POINT_FIVE = 0.4999999;
+
+        void PreciseDecimal::setValue( DecimalType inValue )
+        {
+            if ( inValue < 0 )
             {
-                setValue( inValue );
+                myIsNegative = true;
+                inValue *= -1;
+            }
+            else
+            {
+                myIsNegative = false;
             }
 
+            const auto integer = static_cast<decltype( myInteger )>( inValue );
+            const bool isTooBig = integer > myMaxExpressibleInteger;
 
-            ~impl()
+            if( isTooBig )
             {
-
+                myInteger = myMaxExpressibleInteger;
+                myDecimal = myMaxExpressibleDecimal;
+                return;
             }
 
+            const DecimalType decimalPartAsFloat = inValue - static_cast<DecimalType>( integer );
+            const auto powerMultiplierAsFloat = std::pow( static_cast<DecimalType>( 10 ), static_cast<DecimalType>( getMaxDecimalDigits() ) );
+            const auto powerMultiplier = static_cast<uint64_t>( std::ceil( powerMultiplierAsFloat - POINT_FIVE ) );
+            const auto decimalDigitsShifted = decimalPartAsFloat * static_cast<decltype( decimalPartAsFloat )>( powerMultiplier );
+            const auto decimalDigits = static_cast<decltype( myDecimal )>( std::ceil( decimalDigitsShifted - POINT_FIVE ) );
 
-            impl( const impl& other )
-            : myValue( nullptr )
+            if( decimalDigits > myMaxExpressibleDecimal )
             {
-                if( other.myValue )
+                myInteger = integer + 1;
+                myDecimal = 0;
+                return;
+            }
+
+            myInteger = integer;
+            myDecimal = decimalDigits;
+        }
+
+        short PreciseDecimal::getMaxIntegerDigits() const
+        {
+            return myMaxIntegerDigits;
+        }
+
+        short PreciseDecimal::getMaxDecimalDigits() const
+        {
+            return myMaxDecimalDigits;
+        }
+
+        DecimalType PreciseDecimal::getMaxExpressibleNumber() const
+        {
+            return express( myMaxExpressibleInteger, myMaxExpressibleDecimal, myMaxDecimalDigits, myIsNegative );
+        }
+
+        DecimalType PreciseDecimal::getMinExpressibleNumber() const
+        {
+            return -1 * getMaxExpressibleNumber();
+        }
+
+        std::ostream& PreciseDecimal::toStream( std::ostream& os ) const
+        {
+            if( myInteger == 0 && myDecimal == 0 )
+            {
+                os << "0";
+                return os;
+            }
+
+            if( myIsNegative )
+            {
+                os << "-";
+            }
+
+            os << myInteger;
+
+            if( myDecimal == 0 )
+            {
+                return os;
+            }
+
+            os << ".";
+
+            auto decIntStr = std::to_string( myDecimal );
+            std::stringstream ss;
+
+            bool isNonZeroReached = false;
+
+            for ( auto c = decIntStr.crbegin();
+                  c != decIntStr.crend();
+                  ++c )
+            {
+                if( *c == '0' && !isNonZeroReached )
                 {
-                    myValue = std::make_unique<PreciseDecimal>( *other.myValue );
+                    continue;
+                }
+
+                isNonZeroReached = true;
+                ss << *c;
+            }
+
+            const auto backwards = ss.str();
+
+            short digits = 0;
+            decltype( myDecimal ) tens = 1;
+
+            for( int i = 0; i <= myMaxDecimalDigits; ++i )
+            {
+                ++digits;
+                tens *= 10;
+                const auto r = myDecimal / tens;
+
+                if( r < 1 )
+                {
+                    break;
                 }
             }
 
+            short leadingZeros = myMaxDecimalDigits - digits;
 
-            impl& operator=( const impl& other )
+            for ( short i = 0; i < leadingZeros; ++i )
             {
-                if ( other.myValue )
-                {
-                    myValue = std::make_unique<PreciseDecimal>( *other.myValue );
-                }
-                else
-                {
-                    myValue = nullptr;
-                }
-
-                return *this;
+                os << "0";
             }
 
-            
-            void setValue( DecimalType value )
+            for ( auto c = backwards.crbegin();
+                  c != backwards.crend();
+                  ++c )
             {
-                if( !myValue )
-                {
-                    if( value == 0.0 )
-                    {
-                        return;
-                    }
-                    else
-                    {
-                        allocate();
-                    }
-                }
-                myValue->setValue( value );
+                os << *c;
             }
 
+            return os;
+        }
 
-            DecimalType getValue() const
+        std::string PreciseDecimal::toString() const
+        {
+            std::stringstream ss;
+            toStream( ss );
+            return ss.str();
+        }
+
+        void PreciseDecimal::setMaxIntegerDigits( short inNumDigits )
+        {
+            if( inNumDigits < 1 )
             {
-                if( !myValue )
-                {
-                    return 0.0;
-                }
-                else
-                {
-                    return myValue->getValue();
-                }
+                myMaxIntegerDigits = 1;
+                return;
             }
 
-
-            std::string toString() const
+            if( inNumDigits > MAX_STORAGE_TYPE_DIGITS )
             {
-                if( !myValue )
-                {
-                    return "0";
-                }
-                else
-                {
-                    return myValue->toString();
-                }
+                myMaxIntegerDigits = MAX_STORAGE_TYPE_DIGITS;
+                return;
             }
 
+            myMaxIntegerDigits = inNumDigits;
+        }
 
-            std::ostream& toStream( std::ostream& os ) const
+        void PreciseDecimal::setMaxDecimalDigits( short inNumDigits )
+        {
+            if( inNumDigits < 1 )
             {
-                if( !myValue )
-                {
-                    os << "0";
-                    return os;
-                }
-                else
-                {
-                    return myValue->toStream( os );
-                }
+                myMaxDecimalDigits = 1;
+                return;
             }
 
-
-            void allocate()
+            if( inNumDigits > MAX_STORAGE_TYPE_DIGITS )
             {
-                myValue = std::make_unique<PreciseDecimal>( 19, DEFAULT_PRECISION, 0.0 );
+                myMaxDecimalDigits = MAX_STORAGE_TYPE_DIGITS;
+                return;
             }
 
-        private:
-            std::mutex myMutex;
-            mutable std::unique_ptr<PreciseDecimal> myValue;
-        };
-        
-        
+            myMaxDecimalDigits = inNumDigits;
+        }
+
+        decltype( PreciseDecimal::myInteger ) PreciseDecimal::findMax( short inNumDigits )
+        {
+            decltype( myInteger ) mult = 0;
+            decltype( myInteger ) result = 0;
+
+            for( short i = 0; i < inNumDigits; ++i )
+            {
+                mult = static_cast<decltype( myInteger )>( std::ceil( std::pow( 10, i ) - 0.5) );
+                const auto add = mult * 9;
+                result += add;
+            }
+
+            return result;
+        }
+
+        void PreciseDecimal::setMaxExpressibleInteger()
+        {
+            myMaxExpressibleInteger = findMax( getMaxIntegerDigits() );
+        }
+
+        void PreciseDecimal::setMaxExpressibleDecimal()
+        {
+            myMaxExpressibleDecimal = findMax( getMaxDecimalDigits() );
+        }
+
+        void PreciseDecimal::setMaxExpressibleNumber()
+        {
+            setMaxExpressibleInteger();
+            setMaxExpressibleDecimal();
+        }
+
+        DecimalType PreciseDecimal::express( uint64_t inInteger, uint64_t inDecimal, short inMaxDecimalDigits, bool inIsNegative )
+        {
+            const auto integer = static_cast<DecimalType>( inInteger );
+            const auto decimalPartAsFloat = static_cast<DecimalType>( inDecimal );
+            const auto powerMultiplierAsFloat = std::pow( static_cast<DecimalType>( 10 ), static_cast<DecimalType>( inMaxDecimalDigits ) );
+            const auto powerMultiplier = static_cast<DecimalType>( std::ceil( powerMultiplierAsFloat - 0.5 ) );
+            const auto decimal = decimalPartAsFloat / powerMultiplier;
+            auto result = integer + decimal;
+
+            if( inIsNegative )
+            {
+                result *= -1;
+            }
+
+            return result;
+        }
+
         Decimal::Decimal( DecimalType value )
-        :myImpl( new Decimal::impl( value ) )
+        : myValue{ 19, DEFAULT_PRECISION, value }
         {
 
         }
 
-        
         Decimal::Decimal()
-        :myImpl( new Decimal::impl( 0.0 ) )
+        : Decimal{ 0.0 }
         {
 
         }
-
-
-        Decimal::~Decimal()
-        {
-
-        }
-        
-        Decimal::Decimal( const Decimal& other )
-        :myImpl( new Decimal::impl( *other.myImpl ) )
-        {
-
-        }
-
-
-        Decimal::Decimal( Decimal&& other )
-        :myImpl( std::move( other.myImpl ) )
-        {
-
-        }
-
-
-        Decimal& Decimal::operator=( const Decimal& other )
-        {
-            this->myImpl = std::unique_ptr<Decimal::impl>( new Decimal::impl( *other.myImpl ) );
-            return *this;
-        }
-
-
-        Decimal& Decimal::operator=( Decimal&& other )
-        {
-            myImpl = std::move( other.myImpl );
-            return *this;
-        }
-
 
         DecimalType Decimal::getValue() const
         {
-            return myImpl->getValue();
+            return myValue.getValue();
         }
-
 
         void Decimal::setValue( DecimalType value )
         {
-            myImpl->setValue( value );
+            myValue.setValue( value );
         }
-
 
         void Decimal::parse( const std::string& value )
         {
@@ -193,16 +306,15 @@ namespace mx
             setValue( temp );
         }
 
-
         std::string Decimal::toString() const
         {
-            return myImpl->toString();
+            return myValue.toString();
         }
 
 
         std::ostream& Decimal::toStream( std::ostream& os ) const
         {
-            return myImpl->toStream( os );
+            return myValue.toStream( os );
         }
 
 
@@ -219,214 +331,162 @@ namespace mx
             return value.toStream( os );
         }
 
+        std::string toString( const Decimal& value, unsigned short precision )
+        {
+            MX_UNUSED( precision );
+            return value.toString();
+        }
+
+        std::ostream& toStream( std::ostream& os, const Decimal& value, unsigned short precision )
+        {
+            MX_UNUSED( precision );
+            return value.toStream( os );
+        }
 
         std::ostream& operator<<( std::ostream& os, const Decimal& value )
         {
             return value.toStream( os );
         }
 
-
-        DecimalRange::DecimalRange( DecimalType min, DecimalType max, DecimalType value )
-        :Decimal( value )
-        ,myMin( min )
-        ,myMax( max )
+        DecimalType minExclusive( DecimalType min, DecimalType val )
         {
-            setValue( value );
+            if( val <= min )
+            {
+                return min + NON_ZERO_AMOUNT;
+            }
+
+            return val;
         }
 
-
-        DecimalRange& DecimalRange::operator=( const DecimalRange& other )
+        DecimalType minInclusive( DecimalType min, DecimalType val )
         {
-            setValue( other.getValue() );
-            return *this;
+            if( val < min )
+            {
+                return min;
+            }
+
+            return val;
         }
 
+        DecimalType maxExclusive( DecimalType max, DecimalType val )
+        {
+            if( val >= max )
+            {
+                return max - NON_ZERO_AMOUNT;
+            }
 
-        DecimalRange::DecimalRange( DecimalRange&& other )
-        :Decimal( std::move( other ) )
-        ,myMin( std::move( other.myMin ) )
-        ,myMax( std::move( other.myMax ) )
+            return val;
+        }
+
+        DecimalType maxInclusive( DecimalType max, DecimalType val )
+        {
+            if( val > max )
+            {
+                return max;
+            }
+
+            return val;
+        }
+
+        DecimalType noOp( DecimalType value )
+        {
+            return value;
+        }
+
+        DecimalRange::DecimalRange( DecimalClamp min, DecimalClamp max, DecimalType value )
+        : Decimal{ min( max( value ) ) }
+        , myMinClamp{ min }
+        , myMaxClamp{ max }
         {
 
         }
-
-
-        DecimalRange& DecimalRange::operator=( DecimalRange&& other )
-        {
-            setValue( other.getValue() );
-            return *this;
-        }
-
 
         void DecimalRange::setValue( DecimalType value )
         {
-            if ( value < myMin )
-            {
-                Decimal::setValue( myMin );
-            }
-            else if ( value > myMax )
-            {
-                Decimal::setValue( myMax );
-            }
-            else
-            {
-                Decimal::setValue( value );
-            }
+            Decimal::setValue( myMinClamp( myMaxClamp( value ) ) );
         }
 
-
-        void DecimalRange::parse( const std::string& value )
-        {
-            std::stringstream ss( value );
-            DecimalType temp = 0;
-            if ((ss >> temp).fail() || !(ss >> std::ws).eof())
-            {
-                return;
-            }
-            setValue( temp );
-        }
-        
         PositiveDecimal::PositiveDecimal( DecimalType value )
-        :Decimal( value )
-        {
-            setValue( value );
+        : DecimalRange{
+            std::function<DecimalType(DecimalType)>( []( DecimalType value ){ return minExclusive( 0.0, value ); } ),
+            std::function<DecimalType(DecimalType)>( noOp ),
+            value
         }
+        {
 
+        }
 
         PositiveDecimal::PositiveDecimal()
-        :Decimal( 1 )
+        : PositiveDecimal{ 1 }
         {
 
         }
-        
-
-        void PositiveDecimal::setValue( DecimalType value )
-        {
-            if ( value <= 0 )
-            {
-                Decimal::setValue( NON_ZERO_AMOUNT );
-            }
-            else
-            {
-                Decimal::setValue( value );
-            }
-        }
-
-
-        void PositiveDecimal::parse( const std::string& value )
-        {
-            std::stringstream ss( value );
-            DecimalType temp = NON_ZERO_AMOUNT;
-            if ((ss >> temp).fail() || !(ss >> std::ws).eof())
-            {
-                return;
-            }
-            setValue( temp );
-        }
-
 
         NonNegativeDecimal::NonNegativeDecimal( DecimalType value )
-        :Decimal( value )
-        {
-            setValue( value );
+        : DecimalRange{
+            std::function<DecimalType(DecimalType)>( []( DecimalType value ){ return minInclusive( 0.0, value ); } ),
+            std::function<DecimalType(DecimalType)>( noOp ),
+            value
         }
+        {
 
+        }
 
         NonNegativeDecimal::NonNegativeDecimal()
-        :Decimal( 0 )
+        : NonNegativeDecimal{ 0.0 }
         {
 
-        }
-        
-
-        void NonNegativeDecimal::setValue( DecimalType value )
-        {
-            if ( value < 0 )
-            {
-                Decimal::setValue( 0 );
-            }
-            else
-            {
-                Decimal::setValue( value );
-            }
-        }
-
-
-        void NonNegativeDecimal::parse( const std::string& value )
-        {
-            std::stringstream ss( value );
-            DecimalType temp = 0;
-            if ((ss >> temp).fail() || !(ss >> std::ws).eof())
-            {
-                return;
-            }
-            this->setValue( temp );
         }
 
         
         Percent::Percent( DecimalType value )
-        :DecimalRange( 0, 100, value )
+        : DecimalRange{
+            std::function<DecimalType(DecimalType)>( []( DecimalType value ){ return minInclusive( 0.0, value ); } ),
+            std::function<DecimalType(DecimalType)>( []( DecimalType value ){ return maxInclusive( 100.0, value ); } ),
+            value
+        }
         {
 
         }
 
-
         Percent::Percent()
-        :DecimalRange( 0, 100, 0 )
+        : Percent{ 0.0 }
         {
 
         }
         
-        RotationDegrees::RotationDegrees()
-        :DecimalRange( -180, 180, 0 )
+        RotationDegrees::RotationDegrees( DecimalType value )
+        : DecimalRange{
+            std::function<DecimalType(DecimalType)>( []( DecimalType value ){ return minInclusive( -180.0, value ); } ),
+            std::function<DecimalType(DecimalType)>( []( DecimalType value ){ return maxInclusive( 180.0, value ); } ),
+            value
+        }
         {
 
         }
 
 
-        RotationDegrees::RotationDegrees( DecimalType value )
-        :DecimalRange( -180, 180, value )
+        RotationDegrees::RotationDegrees()
+        : RotationDegrees{ 0.0 }
         {
 
         }
 
         
         TrillBeats::TrillBeats( DecimalType value )
-        :Decimal( value )
+        : DecimalRange{
+            std::function<DecimalType(DecimalType)>( []( DecimalType value ){ return minInclusive( 2.0, value ); } ),
+            std::function<DecimalType(DecimalType)>( noOp ),
+            value
+        }
         {
             this->setValue( value );
         }
 
-
         TrillBeats::TrillBeats()
-        :Decimal( 2 )
+        : TrillBeats{ 0.0 }
         {
 
-        }
-
-
-        void TrillBeats::setValue( DecimalType value )
-        {
-            if ( value < 2 )
-            {
-                Decimal::setValue( 2 );
-            }
-            else
-            {
-                Decimal::setValue( value );
-            }
-        }
-
-        
-        void TrillBeats::parse( const std::string& value )
-        {
-            std::stringstream ss( value );
-            DecimalType temp = 0;
-            if ((ss >> temp).fail() || !(ss >> std::ws).eof())
-            {
-                return;
-            }
-            this->setValue( temp );
         }
     }
 }
